server/pyproject.toml
[project]
name = "quizapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asgiref==3.8.1",
    "certifi==2025.1.31",
    "colorama==0.4.6",
    "distro==1.9.0",
    "django-extensions==4.1",
    "django==5.2",
    "djangorestframework-simplejwt==5.5.0",
    "djangorestframework==3.16.0",
    "h11==0.16.0",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.9.0",
    "openai==1.73.0",
    "pydantic-core==2.33.1",
    "pydantic==2.11.3",
    "pymupdf==1.25.5",
    "python-dotenv==1.1.0",
    "sniffio==1.3.1",
    "sqlparse==0.5.3",
    "tqdm==4.67.1",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "httpcore>=1.0.9",
    "psycopg2-binary>=2.9.10",
    "supabase>=2.15.1",
    "pytz>=2025.2",
    "requests>=2.32.3",
    "django-cors-headers>=4.7.0",
    "celery>=5.5.2",
    "redis>=6.1.0",
    "django-redis>=5.4.0",
    "cryptography>=45.0.2",
]
